requirements.txt
Flask
PyMuPDF
python-docx
requests
beautifulsoup4
transformers
torch
fasttext
langdetect
huggingface_hub
reportlab
pillow
pdfplumber